requirements.txt
boto3>=1.35.0
pillow>=10.0.0